[ codereview ] utils 폴더로 함수 구분
